[Refactor] #4683 alloc-entries.cpp/h からbaseitem-allocation.cpp/h を分離した
