+++ src/Makefile.am
@@ -938,6 +938,7 @@ hengband_SOURCES = \
 	system/terrain-type-definition.cpp system/terrain-type-definition.h \
 	system/gamevalue.h \
 	\
+	system/baseitem/baseitem-allocation.cpp system/baseitem/baseitem-allocation.h \
 	system/baseitem/baseitem-definition.cpp system/baseitem/baseitem-definition.h \
 	system/baseitem/baseitem-key.cpp system/baseitem/baseitem-key.h \
 	system/baseitem/baseitem-list.cpp system/baseitem/baseitem-list.h \
