+++ src/floor/floor-object.cpp
@@ -26,8 +26,8 @@
 #include "object/object-kind-hook.h"
 #include "object/object-stack.h"
 #include "perception/object-perception.h"
-#include "system/alloc-entries.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-allocation.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
