+++ src/system/alloc-entries.h
@@ -8,7 +8,6 @@
 
 #pragma once
 
-#include "util/probability-table.h"
 #include <vector>
 
 enum class MonraceId : short;
@@ -49,60 +48,4 @@ private:
     static MonraceAllocationTable instance;
     MonraceAllocationTable() = default;
     std::vector<MonraceAllocationEntry> entries{};
-};
-
-/*
- * An entry for the object/monster allocation functions
- *
- * Pass 1 is determined from allocation information
- * Pass 2 is determined from allocation restriction
- */
-class BaseitemKey;
-class BaseitemDefinition;
-class BaseitemAllocationEntry {
-public:
-    BaseitemAllocationEntry() = default;
-    BaseitemAllocationEntry(short index, int level, short prob1, short prob2);
-    short index; /* The actual index */
-    int level; /* Base dungeon level */
-    short prob1; /* Probability, pass 1 */
-    short prob2; /* Probability, pass 2 */
-    bool is_same_bi_key(const BaseitemKey &bi_key) const;
-    bool is_chest() const;
-    bool is_gold() const;
-    int get_baseitem_level() const;
-    bool order_level(const BaseitemAllocationEntry &other) const;
-
-private:
-    const BaseitemDefinition &get_baseitem() const;
-    const BaseitemKey &get_bi_key() const;
-};
-
-class BaseitemAllocationTable {
-public:
-    BaseitemAllocationTable(const BaseitemAllocationTable &) = delete;
-    BaseitemAllocationTable(BaseitemAllocationTable &&) = delete;
-    BaseitemAllocationTable operator=(const BaseitemAllocationTable &) = delete;
-    BaseitemAllocationTable operator=(BaseitemAllocationTable &&) = delete;
-    static BaseitemAllocationTable &get_instance();
-
-    void initialize();
-    std::vector<BaseitemAllocationEntry>::iterator begin();
-    std::vector<BaseitemAllocationEntry>::const_iterator begin() const;
-    std::vector<BaseitemAllocationEntry>::iterator end();
-    std::vector<BaseitemAllocationEntry>::const_iterator end() const;
-    size_t size() const;
-    const BaseitemAllocationEntry &get_entry(int index) const;
-    BaseitemAllocationEntry &get_entry(int index);
-    short draw_lottery(int level, uint32_t mode, int count) const;
-    bool order_level(int index1, int index2) const;
-
-    void prepare_allocation();
-
-private:
-    static BaseitemAllocationTable instance;
-    BaseitemAllocationTable() = default;
-    std::vector<BaseitemAllocationEntry> entries;
-
-    ProbabilityTable<int> make_table(int level, uint32_t mode) const;
 };
