+++ src/system/floor-type-definition.cpp
@@ -4,9 +4,9 @@
 #include "game-option/birth-options.h"
 #include "monster/monster-timed-effects.h"
 #include "object-enchant/item-apply-magic.h"
-#include "system/alloc-entries.h"
 #include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-allocation.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
