+++ src/wizard/wizard-item-modifier.cpp
@@ -25,8 +25,8 @@
 #include "object/object-value.h"
 #include "spell-kind/spells-perception.h"
 #include "spell/spells-object.h"
-#include "system/alloc-entries.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-allocation.h"
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/floor-type-definition.h"
